Refactored the code based on my discussion with Gary
