+++ packages/at_client/lib/at_collection/at_json_collection_model.dart
@@ -17,6 +17,8 @@ class AtJsonCollectionModel extends AtCollectionModel {
   Map<String, dynamic> toJson() {
     return jsonModel;
   }
+
+
 }
 
 // Factory for creating instance of AtJsonCollectionModel
