--- packages/at_client/lib/at_collection/model/at_operation_item_status.dart
@@ -1,17 +1,0 @@
-class AtOperationItemStatus {
-  late String atSign;
-  late String key;
-  bool complete;
-  Exception? exception;
-  Operation? operation;
-
-  AtOperationItemStatus({
-    required this.atSign,
-    required this.key,
-    required this.complete,
-    required this.operation,
-    this.exception,
-  });
-}
-
-enum Operation { save, share, unshare, delete }
